feat(widget): moved button disabled
